all done..special characted issue
