+++ Server_folder/sample_file.txt
@@ -1,0 +1,2 @@
+Hi my name is mayank
+This is the project sample file
